Removed verified account code for now Added first stage of the login on connect patches (1/2)
The verified account code will be readded once the second login on connect patch is applied sometime during the development of the next version. If both patches are applied now there will be hell linking the server to servers which do not contain this code, so hense its split in half.