Backported from IRCu 2.10.13. Main changes detailed in Changelog. Changes by me to get it working in Nefarious are in ChangeLog.Nefarious.
1) Implement the concept of a "delayed" /die or /restart.

If you type "/restart", your server restarts normally.  However, you can
also give it a time, i.e. "/restart 300", causing it to wait 300 seconds
before restarting, sending warning messages to all your users a la UNIX
"shutdown" command.  You can cancel a pending restart with "/restart
cancel".  You can even give a reason: "/restart 300 :Upgrading to new
server version" (note the ':').  Both the time and the reason are optional,
and the only restrictions are that the reason cannot start off with numbers
or be the word "cancel" if you omit the time :)

2) Implement restrictions on new connections.

If you have a delayed /die or /restart pending, new connections will be
rejected with a notice that the server is restarting.  You can make certain
ports exempt from this restriction by adding an X flag to the p:line--
you may want to do this on any ports specifically for your opers, so
that you can get in and stop a pending /die or /restart if necessary.

3) Implement network-triggered /die, and /restart. (rehash already added)

NOTE: These features default to disabled.  They are mostly intended for
centrally managed IRC networks...

It is now possible to issue a /die, or /restart remotely (but only
from a pseudo-server connection, such as x3).  These commands may be
issued either to specific servers or to every network server.  In all
cases, the first argument must be either "*" or the numeric of the target
server. For /restart and /die, the delay time and reason must be
specified...even if the delay time is set to "cancel".  Note that using a
delay time of 0 may not do what you want in the case of the entire network;
the message may not get propagated to every server due to how buffering in
ircu works.

This feature is under control of the NETWORK_REHASH, NETWORK_DIE, and
NETWORK_RESTART features, and defaults to FALSE.  Make sure you know what
you're doing before enabling these features.